refactor(lib): delete the Graph Access Toolkit - it was all shims and dead SQL
`packages/lib/src/shared/surreal.ts` said of itself: "This whole file goes in
`c-surreal-delete`." Twenty exports, and by the time this sweep reached it, ten
had no caller at all and the rest were prose mentions in comments. The single
real runtime import chain ran `json.ts -> shared/surql.ts -> surreal.ts` for
`surrealLiteral`, an alias with ZERO call sites.

What was actually load-bearing had already moved: `safeKeyPart`,
`recordKeyPart`, `isoTimestamp` and the typed row-field readers live in
`shared/row-fields.ts` and are engine-neutral - DuckDB row ids are derived
exactly as SurrealDB record ids were. The toolkit only re-exported them, and
`derive-keys.ts` / `record-select.ts` re-exported the toolkit, so the shim chain
pointed in a circle.

Deleted: `shared/surreal.ts`, `shared/surql.ts`, `shared/record-select.ts`, and
`json.ts`'s `surrealLiteral`. `derive-keys.ts` and `metrics/util.ts` now
re-export straight from `row-fields.ts` - both keep their names because ~30 call
sites say what they want by importing them.

`surreal.test.ts` becomes `row-fields.test.ts`, keeping every test for the
helpers that survive (row-field access + key derivation, 278 passing) and
dropping only the sections that tested SurrealQL literal encoders.

`scripts/check-record-select.ts` stays wired but is now honest in its header: it
guards a shape no module can produce any more, and it is kept solely because
ci.yml runs it and ci.yml cannot be pushed from this branch. Its exclusion list
is empty, so any future hit is unambiguously a mistake.
