+++ packages/lib/src/json.ts
@@ -7,19 +7,11 @@
  *    encode is not the right tool here; the CLI just wants a 2-space-indented
  *    dump of a value it already trusts (computed in-process, no external
  *    payload).
- *  - `surrealLiteral` quotes a string for inclusion inside a SurrealQL
- *    statement literal. It is a thin re-export of `surrealString` from the
- *    shared SurrealQL write seam (`lib/shared/surql.ts`), kept here for its
- *    existing call sites; new code should import `surrealString` directly.
  *
  * Real decode boundaries (`JSON.parse` on jsonl lines / file payloads) live in
  * `src/ingest/*` and should use Effect Schema decoders directly - those are
  * tracked in issue #86 ("Effect JSON-boundary Schema decoders").
  */
 
-import { surrealString } from "./shared/surql.ts";
-
 export const prettyPrint = (value: unknown): string =>
     JSON.stringify(value, null, 2);
-
-export const surrealLiteral = surrealString;
